\name{loadPhenotypes}
\alias{loadPhenotypes}
\title{
Load phenotypes from phenotype file
}
\description{
Load all or part of phenotypes from phenotype file based on user option.
}
\usage{
loadPhenotypes(opt)
}
\arguments{
  \item{opt}{The list of input options provided by user.}
}
\details{
Load all or part of phenotypes from phenotype file based on user option.
}
\value{
Return a data frame object that conbtains all the required phenotypes. 
}

\examples{
#phenotype = loadPhenotypes(opt)
}

