\name{PHESANT-package}
\alias{PHESANT-package}
\alias{PHESANT}
\docType{package}
\title{
\packageTitle{PHESANT}
}
\description{
\packageDescription{PHESANT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PHESANT}
\packageIndices{PHESANT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PHESANT}

Maintainer: \packageMaintainer{PHESANT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
