\name{run}
\alias{run}
\title{
Run an PHESANT analysis with user specificed options
}
\description{
Run an PHESANT analysis with user specificed options
}
\usage{
run(opt)
}
\arguments{
  \item{opt}{The list of input options provided by user.}
}

\value{
No return value. All output will be writen to the user specificed oiutput folder.
}

\examples{
# run(opt)
}

