\name{PHESANT-package}
\alias{PHESANT-package}
\alias{PHESANT}
\docType{package}
\title{
\packageTitle{PHESANT}
}
\description{
\packageDescription{PHESANT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PHESANT}
\packageIndices{PHESANT}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{PHESANT}

Maintainer: \packageMaintainer{PHESANT}
}
\references{
Literature or other references for background information 
}
\keyword{ package }
\seealso{
To be done
}
\examples{
#example here 
}
